requirements.txt
# Flask Framework
Flask>=1.1.2
Flask-SocketIO>=4.3.0
sqlalchemy>=1.3.17

bcrypt>=3.1.7
pyjwt>=1.7.1